Use eeprom pin info within rboot and otbiot
Specifically now reads and uses:
- soft reset pin (rboot and otbiot)
- status led pin (otbiot only)

If eeprom not present or invalid info defaults to:
- no soft reset pin
- no status led pin

Default is wemos d1 mini hardware.

Subsequent commit will add support for otbiot v0.3 hardware with compile flag

+++ include/otb_led.h
@@ -226,7 +226,7 @@ void otb_led_control_on_timer(void *arg);
 void otb_led_control_init(otb_led_sequence *seq);
 bool otb_led_control_seq(otb_led_sequence *seq);
 void otb_led_wifi_update(uint32_t rgb, bool store);
-void otb_led_neo_update(uint32_t *rgb, int num, uint8_t pin);
+void otb_led_neo_update(uint32_t *rgb, int num, uint32_t pin);
 uint32_t otb_led_neo_get_rgb(uint8_t red, uint8_t green, uint8_t blue);
 void otb_led_wifi_blink(uint8 times);
 void otb_led_wifi_init_blink_timer(void);
